+++ src/app/robots.ts
@@ -1,0 +1,12 @@
+import type { MetadataRoute } from 'next';
+import { getSiteUrl } from '@/utils/seo';
+
+export default function robots(): MetadataRoute.Robots {
+  const base = getSiteUrl();
+  return {
+    rules: [{ userAgent: '*', allow: '/' }],
+    sitemap: `${base}/sitemap.xml`,
+    host: base,
+  };
+}
+
