+++ src/utils/seo.ts
@@ -1,0 +1,23 @@
+export function truncate(text: string, max: number): string {
+  if (!text) return '';
+  if (text.length <= max) return text;
+  return text.slice(0, Math.max(0, max - 1)).trimEnd() + '…';
+}
+
+export function getSiteUrl(): string {
+  const base = process.env.NEXT_PUBLIC_SITE_URL || 'http://localhost:3000';
+  try {
+    return new URL(base).toString().replace(/\/$/, '');
+  } catch {
+    return 'http://localhost:3000';
+  }
+}
+
+export function toAbsoluteUrl(pathOrUrl: string): string {
+  if (!pathOrUrl) return getSiteUrl();
+  if (/^https?:\/\//i.test(pathOrUrl)) return pathOrUrl;
+  const base = getSiteUrl();
+  const path = pathOrUrl.startsWith('/') ? pathOrUrl : `/${pathOrUrl}`;
+  return `${base}${path}`;
+}
+
